🚐 move redux file structure
